# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Mock сервис для имитации API Центрального Банка России",
        "title": "MOCK-CBR API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/health": {
            "get": {
                "description": "Returns status OK",
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/scripts/XML_daily.asp": {
            "get": {
                "description": "Returns XML with currency rates for given date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/xml"
                ],
                "tags": [
                    "rates"
                ],
                "summary": "Get currency rates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in DD/MM/YYYY",
                        "name": "date_req",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "XML data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid date format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "data not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}